Change action updateasssets - build ref from dev
Changing the buildref file to be pulled from development branch now so we no longer keep that updated based on releases. Dev branch is kept current as well, and no difference pulling that from dev or master the way we manage it. **Just need to be more vigilant on PRs**